move sdk methods back into SDK class
